nutri-ai-backend/requirements.txt
# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Grounding DINO (vision-language, zero-shot)
transformers>=4.44.0
torch>=2.0.0
torchvision>=0.15.0

# Imágenes
Pillow>=10.0.0
python-multipart>=0.0.6
